Proposed change to resolve import cycle
