+++ cumulusci/core/tasks.py
@@ -8,7 +8,9 @@ import os
 import time
 import threading
 
+from cumulusci import __version__
 from cumulusci.utils import cd
+from cumulusci.core.exceptions import ServiceNotValid, ServiceNotConfigured
 from cumulusci.core.exceptions import TaskRequiresSalesforceOrg
 from cumulusci.core.exceptions import TaskOptionsError
 
@@ -223,3 +225,27 @@ class BaseTask(object):
             }
         )
         return [ui_step]
+
+
+class BaseSalesforceTask(BaseTask):
+    """Base for tasks that need a Salesforce org"""
+
+    name = "BaseSalesforceTask"
+    salesforce_task = True
+
+    def _get_client_name(self):
+        try:
+            app = self.project_config.keychain.get_service("connectedapp")
+            return app.client_id
+        except (ServiceNotValid, ServiceNotConfigured):
+            return "CumulusCI/{}".format(__version__)
+
+    def _run_task(self):
+        raise NotImplementedError("Subclasses should provide their own implementation")
+
+    def _update_credentials(self):
+        orig_config = self.org_config.config.copy()
+        self.org_config.refresh_oauth_token(self.project_config.keychain)
+        if self.org_config.config != orig_config:
+            self.logger.info("Org info updated, writing to keychain")
+            self.project_config.keychain.set_org(self.org_config)
